+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Improve subscription & state update performance (#325) - @mjarvis
 - Enable build settings "Allow app extension API only" (#328) - @oradyvan
 - Open `Subscription<State>` to allow external extensions (#383) - @mjarvis
+- Update project to Swift 4.2 (#256, #335, #374) - @mjarvis, @DivineDominion
 
 # 4.0.1